api/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
--index-url http://build.locusdev.net:10001/simple/
--extra-index-url https://pypi.python.org/simple/
--trusted-host build.locusdev.net

amqp==2.5.2               # via kombu
appdirs==1.4.3            # via black, virtualenv
argh==0.26.2              # via watchdog
asgiref==3.2.7            # via django
astroid==2.3.3            # via pylint
attrs==19.3.0             # via black, pytest
billiard==3.6.3.0         # via celery
black==19.10b0            # via -r requirements.in
celery==4.4.2             # via -r requirements.in, django-celery-results
cfgv==3.1.0               # via pre-commit
click==7.1.1              # via black, pip-tools
distlib==0.3.0            # via virtualenv
django-celery-results==1.2.1  # via -r requirements.in
django-extensions==2.2.9  # via -r requirements.in
django==3.0.5             # via -r requirements.in, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
factory-boy==2.12.0       # via -r requirements.in
faker==4.0.3              # via factory-boy
filelock==3.0.12          # via virtualenv
identify==1.4.14          # via pre-commit
isort==4.3.21             # via pylint
kombu==4.6.8              # via celery
lazy-object-proxy==1.4.3  # via astroid
mccabe==0.6.1             # via pylint
more-itertools==8.2.0     # via pytest
mypy-extensions==0.4.3    # via mypy
mypy==0.770               # via -r requirements.in
nodeenv==1.3.5            # via pre-commit
packaging==20.3           # via pytest
pathspec==0.8.0           # via black
pathtools==0.1.2          # via watchdog
pip-tools==5.0.0          # via -r requirements.in
pluggy==0.13.1            # via pytest
pre-commit==2.2.0         # via -r requirements.in
py==1.8.1                 # via pytest
pylint==2.4.4             # via -r requirements.in
pyparsing==2.4.7          # via packaging
pytest-django==3.9.0      # via -r requirements.in
pytest==5.4.1             # via -r requirements.in, pytest-django
python-dateutil==2.8.1    # via faker
pytz==2019.3              # via celery, django
pyyaml==5.3.1             # via pre-commit, watchdog
regex==2020.4.4           # via black
six==1.14.0               # via astroid, django-extensions, packaging, pip-tools, python-dateutil, virtualenv
sqlparse==0.3.1           # via django
text-unidecode==1.3       # via faker
toml==0.10.0              # via black, pre-commit
typed-ast==1.4.1          # via black, mypy
typing-extensions==3.7.4.2  # via mypy
vine==1.3.0               # via amqp, celery
virtualenv==20.0.18       # via pre-commit
watchdog[watchmedo]==0.10.2  # via -r requirements.in
wcwidth==0.1.9            # via pytest
wrapt==1.11.2             # via astroid

# The following packages are considered to be unsafe in a requirements file:
# pip
